Urdu-chatbot/requirements.txt
streamlit
streamlit-mic-recorder
langchain-google-genai
gtts
langchain
PyPDF2 
python-docx  